+++ src/components/films/films-list.js
@@ -1,11 +1,6 @@
 import Abstract from "../abstract";
 
 export default class FilmsList extends Abstract {
-
-  // getFilmsList() {
-  //   return this.getElement().querySelector(`.films-list`);
-  // }
-
   getFilmsListContainer() {
     return this.getElement().querySelector(`.films-list__container`);
   }
