+++ src/components/films/films.js
@@ -1,6 +1,11 @@
 import Abstract from "../abstract";
 
 export default class Films extends Abstract {
+  getTemplate() {
+    return (
+      `<section class="films"></section>`
+    );
+  }
 
   getFilmsList() {
     return this.getElement().querySelector(`.films-list`);
@@ -8,11 +13,5 @@ export default class Films extends Abstract {
 
   getFilmsListContainer() {
     return this.getElement().querySelector(`.films-list__container`);
-  }
-
-  getTemplate() {
-    return (
-      `<section class="films"></section>`
-    );
   }
 }
